+++ frontend/src/scenes/onboarding/sdks/error-tracking/ErrorTrackingSDKInstructions.tsx
@@ -17,14 +17,15 @@ import {
     RubyOnRailsInstallation,
     SvelteInstallation,
 } from '@posthog/shared-onboarding/error-tracking'
-import { JSEventCapture, PythonEventCapture } from '@posthog/shared-onboarding/product-analytics'
+import { PythonEventCapture } from '@posthog/shared-onboarding/product-analytics'
 
 import { SDKInstructionsMap, SDKKey } from '~/types'
 
+import { JS_WEB_SNIPPETS as BASE_JS_WEB_SNIPPETS } from '../shared/jsWebSnippets'
 import { withOnboardingDocsWrapper } from '../shared/onboardingWrappers'
 
 const JS_WEB_SNIPPETS = {
-    JSEventCapture,
+    ...BASE_JS_WEB_SNIPPETS,
 }
 
 const PYTHON_SNIPPETS = {
