+++ frontend/src/scenes/onboarding/sdks/session-replay/SessionReplaySDKInstructions.tsx
@@ -27,9 +27,11 @@ import { urls } from 'scenes/urls'
 
 import { OnboardingStepKey, SDKInstructionsMap, SDKKey } from '~/types'
 
+import { JS_WEB_SNIPPETS } from '../shared/jsWebSnippets'
 import { withOnboardingDocsWrapper } from '../shared/onboardingWrappers'
 
 const SNIPPETS = {
+    ...JS_WEB_SNIPPETS,
     SessionReplayFinalSteps,
 }
