+++ frontend/src/scenes/onboarding/sdks/workflows/WorkflowsSDKInstructions.tsx
@@ -41,9 +41,11 @@ import {
     WebInstallation,
     ZapierInstallation,
 } from '@posthog/shared-onboarding/product-analytics'
+import { NodeEventCapture, PythonEventCapture } from '@posthog/shared-onboarding/product-analytics'
 import { StepDefinition } from '@posthog/shared-onboarding/steps'
 
 import { useMDXComponents } from 'scenes/onboarding/OnboardingDocsContentWrapper'
+import { JS_WEB_SNIPPETS } from 'scenes/onboarding/sdks/shared/jsWebSnippets'
 
 import { SDKInstructionsMap, SDKKey, SDKTag, SDKTagOverrides } from '~/types'
 
@@ -76,10 +78,14 @@ function workflowsModifySteps(steps: StepDefinition[]): StepDefinition[] {
     ]
 }
 
+const NODE_SNIPPETS = { NodeEventCapture }
+const PYTHON_SNIPPETS = { PythonEventCapture }
+
 // JS Web SDKs
 const WorkflowsWebInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: WebInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 
 // Frontend frameworks
@@ -87,80 +93,97 @@ const WorkflowsReactInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: ReactInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'React',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsNextJSInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: NextJSInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Next.js',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsSvelteInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: SvelteInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Svelte',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsAstroInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: AstroInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Astro',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsTanStackInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: TanStackInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'TanStack Start',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsAngularInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: AngularInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Angular',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsVueInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: VueInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Vue',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsNuxtJSInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: NuxtInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Nuxt',
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsReactRouterInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: ReactRouterInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsRemixJSInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: RemixInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'React Router',
+    snippets: JS_WEB_SNIPPETS,
 })
 
 // Website builders
 const WorkflowsBubbleInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: BubbleInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsFramerInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: FramerInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsWebflowInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: WebflowInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsDocusaurusInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: DocusaurusInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsGoogleTagManagerInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: GoogleTagManagerInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsShopifyInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: ShopifyInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsWordpressInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: WordpressInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: JS_WEB_SNIPPETS,
 })
 const WorkflowsRetoolInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: RetoolInstallation,
@@ -192,16 +215,19 @@ const WorkflowsRNInstructionsWrapper = withOnboardingDocsWrapper({
 const WorkflowsNodeInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: NodeJSInstallation,
     modifySteps: workflowsModifySteps,
+    snippets: NODE_SNIPPETS,
 })
 const WorkflowsPythonInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: PythonInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Python',
+    snippets: PYTHON_SNIPPETS,
 })
 const WorkflowsDjangoInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: DjangoInstallation,
     modifySteps: workflowsModifySteps,
     wizardIntegrationName: 'Django',
+    snippets: PYTHON_SNIPPETS,
 })
 const WorkflowsGoInstructionsWrapper = withOnboardingDocsWrapper({
     Installation: GoInstallation,
