+++ frontend/src/scenes/settings/environment/SDKSetupInstructions.tsx
@@ -16,9 +16,6 @@ import {
     GoInstallation,
     GoogleTagManagerInstallation,
     IOSInstallation,
-    JSEventCapture,
-    JSHtmlSnippet,
-    JSInitSnippet,
     LaravelInstallation,
     NextJSInstallation,
     NodeEventCapture,
@@ -41,29 +38,13 @@ import {
 } from '@posthog/shared-onboarding/product-analytics'
 import type { StepDefinition } from '@posthog/shared-onboarding/steps'
 
-import { useJsSnippetConfig } from 'lib/components/JSSnippet'
 import { Link } from 'lib/lemon-ui/Link'
 import { OnboardingDocsContentWrapper } from 'scenes/onboarding/OnboardingDocsContentWrapper'
 import SetupWizardBanner from 'scenes/onboarding/sdks/sdk-install-instructions/components/SetupWizardBanner'
+import { JS_WEB_SNIPPETS } from 'scenes/onboarding/sdks/shared/jsWebSnippets'
 import { teamLogic } from 'scenes/teamLogic'
 
-import { SDK_DEFAULTS_DATE } from '~/loadPostHogJS'
 import { SDKKey } from '~/types'
-
-const InAppJSHtmlSnippet = (): JSX.Element => {
-    const config = useJsSnippetConfig()
-    return <JSHtmlSnippet {...config} />
-}
-
-const InAppJSInitSnippet = (): JSX.Element => {
-    return <JSInitSnippet defaultsDate={SDK_DEFAULTS_DATE} />
-}
-
-const JS_WEB_SNIPPETS = {
-    JSEventCapture,
-    JSHtmlSnippet: InAppJSHtmlSnippet,
-    JSInitSnippet: InAppJSInitSnippet,
-}
 const NODE_SNIPPETS = { NodeEventCapture }
 const PYTHON_SNIPPETS = { PythonEventCapture }
